+++ mapfile.h
@@ -1,2 +1,2 @@
-void * mapFile(const char * path, int openflags, int mmapprot, int mmapflags, int * length);
+void * mapFile(const char * path, int openflags, int mmapprot, int mmapflags, size_t * length);
